+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
   -------------------------------------------------------------------
 ## [Unreleased]
+
+## [1.8.0] 2020-01-16
 - Added `--dry-run` option to run all the non-destructive fail-fast options without fear of actual process execution.
 - Fixed a bug in mysql provider where using the default unspecified port would cause an error.
 - Fixed a bug in postgres provider where seed_rows was being ignored.